made check_objects file to be able to test negative values in methods for many classes, updated code in Habit and Species to make unittest passed. started on updates to example_input_params to ajdust for new changes
